api/requirements.txt
asgiref==3.3.4
cryptography==3.4.8
dj-database-url==0.5.0
dj-rest-auth==2.1.9
Django==3.2.4
django-allauth==0.50.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
freezegun==1.5.1
gunicorn==20.1.0
oauthlib==3.1.1
Pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.15
whitenoise==6.4.0
